requirements.txt
csscompressor==0.9.5
libsass==0.19.2
liquidpy==0.0.7
Markdown==3.1.1
Pygments==2.4.2
python-dateutil==2.8.0
PyYAML==5.1.2
six==1.12.0
